(GH-87) Moved to using reduced output
The old implementation caused promotional message, and chocolatey banner to
get mixed into the works. If upgrading Chocolatey, cause problems as the
version caught is a string with "vX.X.X" and in the where clause is attempted
to be converted to a .NET version type.

Limiting the output reduces it to only the packages, leaving out any banners.